+++ react/FormattedPrice.tsx
@@ -1,4 +1,4 @@
-import React, { FC, Fragment } from 'react'
+import React, { FC } from 'react'
 import { injectIntl, InjectedIntlProps, FormattedMessage } from 'react-intl'
 import { FormattedCurrency } from 'vtex.format-currency'
 
@@ -6,7 +6,7 @@ const FormattedPrice: FC<FormattedPriceProps & InjectedIntlProps> = ({
   value,
 }) => {
   return (
-    <Fragment>
+    <div className={value === 0 ? 'c-success' : ''}>
       {value === 0 ? (
         <FormattedMessage id="store/price.Free" />
       ) : value === null || value === undefined ? (
@@ -14,7 +14,7 @@ const FormattedPrice: FC<FormattedPriceProps & InjectedIntlProps> = ({
       ) : (
         <FormattedCurrency value={value} />
       )}
-    </Fragment>
+    </div>
   )
 }