+++ CHANGELOG.md
@@ -7,15 +7,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `c-success` token to component when the received value is zero.
+
 ## [0.2.0] - 2019-09-19
 
-## Added
+### Added
 
 - Show TBA in case of undefined or null price.
 
 ## [0.1.0] - 2019-08-28
 
-## Added
+### Added
 
 - initial setup